move loginWithDialog to authService
